+++ src/elements/Button/DisruptiveButton.jsx
@@ -25,6 +25,7 @@ const Page = React.createClass({
 ```
 */
 const DisruptiveButton = React.createClass({
+    displayName: 'DisruptiveButton',
     // template rendering
     render() {
         // split 'normal' props from button content