+++ src/elements/Nothing/Nothing.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
 const Nothing = React.createClass({
+    displayName: 'Nothing',
     // render nothing
     render() {
         return false;